feat : hpUI, fix : warning message

+++ Jazz Jack Rabbit 2/Image.cpp
@@ -163,8 +163,8 @@ void Image::Render(HDC hdc, int destX, int destY, int frameX, int frameY, float 
 			hdc,
 			destX - (imageInfo->frameWidth / 2 ),
 			destY - (imageInfo->frameHeight),
-			imageInfo->frameWidth * scale,
-			imageInfo->frameHeight * scale,	// 전체 프레임 수
+			(int)imageInfo->frameWidth * scale,
+			(int)imageInfo->frameHeight * scale,	// 전체 프레임 수
 			imageInfo->hMemDc,
 			imageInfo->frameWidth * frameX,
 			imageInfo->frameHeight * frameY,
